csvtobigquery/requirements.txt
google-cloud-bigquery==3.3.1

